requirements.txt
moviepy==1.0.3
scipy==1.10.1
numpy==1.24.3
openai-whisper
ffmpeg-python==0.2.0